pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fhad"
version = "0.3.0"
description = "Production baseline + web app for fetal head abnormalities detection"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1",
  "numpy>=1.24",
  "Pillow>=10.0",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "python-multipart>=0.0.9",
  "jinja2>=3.1"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
